reorganized components folder.
