+++ src/view/components/index.js
@@ -1,20 +1,21 @@
 //this exports Nav from Nav.js
-import Nav from "./Nav";
-import Chef from "./Chef";
-import Home from "./Home";
-import About from "./About";
-import Recipes from "./Item";
-import Cocktail from "./Cocktail";
-import Nutrition from "./Nutrition";
-import RecipeCards from "./RecipeCards";
+import _Nav from "./Nav";
+import _Home from "./Home";
+import _ItemCards from "./items/ItemCards";
+import _Item from "./items/Item";
+import _LogIn from "./forms/LogIn";
+import _SignUp from "./forms/SignUp";
+import _Owner from "./Owner";
+import _User from "./User"
+
 
 export {
-  Nav as AppNav,
-  Chef as AppChef,
-  Home as AppHome,
-  About as AppAbout,
-  Recipes as AppRecipes,
-  Cocktail as AppCocktail,
-  Nutrition as AppNutrition,
-  RecipeCards as AppRecipeCards
+  _Nav as Nav,
+  _Home as Home,
+  _ItemCards as ItemCards,
+  _Item as Item,
+  _LogIn as LogIn,
+  _SignUp as SignUp,
+  _Owner as Owner,
+  _User as User
 };
