+++ src/view/components/items/ItemCards.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { connect } from "react-redux";
-import { getRecipe } from "../../state/_shared/middleware/api";
+import { getRecipe } from "../../../state/_shared/middleware/api";
 import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardActionArea from "@material-ui/core/CardActionArea";
